split registers into a seperate class, moved common functionality from assembly dialects to backend, added test
