+++ tests/dialects/test_x86.py
@@ -1,0 +1,7 @@
+from xdsl.dialects import x86
+
+
+def test_register():
+    rax = x86.Registers.RAX
+    assert rax.is_allocated
+    assert rax.register_name == "rax"
